Add files via upload
The first part of this code evaluates the inverse of a square matrix (if it exists) using Gaussian row-elimination. The first part's code is then used with a few modifications in the second part to find the multiplicative inverse of a square matrix (if it exists) over the field Z_257. For this, we take the first part's code and replace all division steps by modular division (i.e. a/b becomes a * mod_inv(b), where mod_inv(b) is the multiplicative inverse of b mod 257).